ollama/requirements.txt
google-generativeai
python-dotenv
nltk